@@ -1,23 +1,30 @@
 {
   "name": "japanpost-zipcode",
   "description": "Japan Post Zipcode",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "author": "@kawanet",
   "bugs": {
     "url": "https://github.com/kawanet/japanpost-zipcode/issues"
   },
   "dependencies": {
-    "axios": "^0.18.0",
-    "iconv-lite": "^0.4.24",
-    "jszip": "^3.1.5"
+    "axios": "^0.21.1",
+    "iconv-lite": "^0.6.2",
+    "jszip": "^3.6.0"
   },
   "devDependencies": {
-    "@types/jszip": "^3.1.4",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^10.11.3",
-    "mocha": "^5.2.0",
-    "typescript": "^3.1.1"
+    "@types/jszip": "^3.4.0",
+    "@types/mocha": "^8.2.2",
+    "@types/node": "^14.14.44",
+    "mocha": "^8.3.2",
+    "typescript": "^4.2.4"
   },
+  "files": [
+    "LICENSE",
+    "README.md",
+    "lib/*.js",
+    "tmp/.gitignore",
+    "types/*.d.ts"
+  ],
   "homepage": "https://github.com/kawanet/japanpost-zipcode#readme",
   "keywords": [
     "ajaxzip",
@@ -32,10 +39,10 @@
     "url": "git+https://github.com/kawanet/japanpost-zipcode.git"
   },
   "scripts": {
+    "build": "./node_modules/.bin/tsc",
     "fixpack": "fixpack",
+    "prepack": "npm run build && npm test",
     "test": "./node_modules/.bin/mocha test",
-    "tsc": "./node_modules/.bin/tsc",
-    "tsc-types": "./node_modules/.bin/tsc -p types/types.json && perl -i -pe 's/^ *private .*\\n//' types/ken_all.d.ts",
     "tsc-watch": "./node_modules/.bin/tsc --watch",
     "update": "/bin/rm -f tmp/ken_all.* && node -e \"require('./').KenAll.readAll({logger:console})\""
   },
